Quiz: Amati: Phase 2: Handle non-UTF8 uploaded responses #135778
Uploaded responses need to be in utf8 format. Authors using Excel will
often have xA0 or #160 (non-breaking spaces) present in the data, as
well as other extended characters allowed by iso-8859-1 encoding.
Therefore, if the file is not valid utf8, we assume it is iso-8859-1
and convert it to UTF8.